Add handle function for GET requests in URL shortener
A new function, 'HandleGet', has been added to the handlers package to manage GET requests in the URL shortener command. This function retrieves the specific URL associated with the received ID and redirects to it. It also considers cases when ID is empty or if the URL is not found in the store.