Pass configuration to HandlePost in server.go
Extended HandlePost function's parameters in server.go to include the 'config' apart from 'store'. This resulted due to the removal of line creating a new FlagConfig in post.go file, instead opting to pass it in from the server.go file. This helps to maintain consistency in configuration across different parts of the code.